multi upload front

+++ src/main/java/pl/coderslab/rentier/controller/admin/ProductController.java
@@ -2,6 +2,8 @@ package pl.coderslab.rentier.controller.admin;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -10,16 +12,18 @@ import org.springframework.web.multipart.MultipartFile;
 import pl.coderslab.rentier.beans.ProductSearch;
 import pl.coderslab.rentier.entity.*;
 import pl.coderslab.rentier.exception.InvalidFileException;
+import pl.coderslab.rentier.pojo.FileImport;
 import pl.coderslab.rentier.repository.BrandRepository;
 import pl.coderslab.rentier.repository.ProductCategoryRepository;
+import pl.coderslab.rentier.repository.ProductImageRepository;
 import pl.coderslab.rentier.repository.ProductRepository;
 import pl.coderslab.rentier.service.ImageServiceImpl;
 import pl.coderslab.rentier.service.ProductServiceImpl;
 
 import javax.validation.Valid;
-import java.io.File;
 import java.io.IOException;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 
 
@@ -27,20 +31,27 @@ import java.util.List;
 @RequestMapping("/admin/products")
 public class ProductController {
 
+    private final org.slf4j.Logger logger = LoggerFactory.getLogger(ProductController.class);
+
+    @Value("${rentier.productMaxImagesCount}")
+    private int productMaxImagesCount;
+
     private final ProductRepository productRepository;
     private final ProductCategoryRepository productCategoryRepository;
     private final BrandRepository brandRepository;
     private final ProductServiceImpl productService;
     private final ImageServiceImpl imageService;
+    private final ProductImageRepository productImageRepository;
 
     public ProductController(ProductRepository productRepository,
                              ProductCategoryRepository productCategoryRepository, BrandRepository brandRepository,
-                             ProductServiceImpl productService, ImageServiceImpl imageService) {
+                             ProductServiceImpl productService, ImageServiceImpl imageService, ProductImageRepository productImageRepository) {
         this.productRepository = productRepository;
         this.productCategoryRepository = productCategoryRepository;
         this.brandRepository = brandRepository;
         this.productService = productService;
         this.imageService = imageService;
+        this.productImageRepository = productImageRepository;
     }
 
 
@@ -100,7 +111,8 @@ public class ProductController {
 
     @PostMapping("/form")
     public String saveProduct(Model model, @ModelAttribute @Valid Product product, BindingResult resultProduct,
-                              @RequestParam(value = "file", required = false) MultipartFile file) {
+                              @RequestParam(value = "file", required = false) MultipartFile file,
+                              @RequestParam(value = "files", required = false) MultipartFile[] files) {
 
 
         if (product.getId() == null) {
@@ -118,17 +130,50 @@ public class ProductController {
 
         }
 
+        if (files.length > productMaxImagesCount) {
+            resultProduct.rejectValue("productName", "error.files", "Możesz dodać maksymalnie " + productMaxImagesCount + " plików");
+        }
+
         String savedFileName = null;
         if (!"".equals(file.getOriginalFilename())) {
             try {
                 savedFileName = productService.saveProductImage(file, product);
+                product.setImageFileName(savedFileName);
             } catch (InvalidFileException e) {
                 resultProduct.rejectValue("productName", "error.fileName", "Niepoprawny plik");
             } catch (IOException e) {
                 resultProduct.rejectValue("productName", "error.fileName", "Błąd odczytu/zapisu plik");
             }
 
+        } else {
+
+            resultProduct.rejectValue("productName", "error.fileName", "Musisz wybrać zdjęcie główne");
+
         }
+
+        List<ProductImage> savedImages = new ArrayList<>();
+        for (MultipartFile imageFile: files) {
+            try {
+                savedFileName = productService.saveProductImage(imageFile, product);
+                ProductImage productImage = new ProductImage();
+                productImage.setImageFileName(savedFileName);
+                productImage.setProduct(product);
+                productImage.setMain_image(false);
+                savedImages.add(productImage);
+                logger.info("Saved image " + savedFileName);
+            } catch (InvalidFileException e) {
+                resultProduct.rejectValue("productImages", "error.fileName", "Niepoprawny plik " + imageFile.getOriginalFilename());
+            } catch (IOException e) {
+                resultProduct.rejectValue("productImages", "error.fileName", "Błąd odczytu/zapisu pliku " + imageFile.getOriginalFilename());
+            }
+        }
+
+
+        if (savedImages.size() > 0) {
+            logger.info("Saving productImages");
+            product.setProductImages(savedImages);
+        }
+
 
         if (product.getId() != null && "".equals(file.getOriginalFilename())) {
             product.setImageFileName(productRepository.selectImageFileNameByProductId(product.getId()));
@@ -136,13 +181,26 @@ public class ProductController {
 
 
         if (resultProduct.hasErrors()) {
+            product.setProductImages(new ArrayList<>());
+            product.setImageFileName(null);
+            if (product.getImageFileName() != null) {
+                logger.info("savedFileName: " + savedFileName);
+                productService.deleteProductImage(savedFileName);
+            }
+            if (savedImages.size()>0) {
+                logger.info("Images to delete: " + savedImages.toString());
+                savedImages.stream()
+                        .forEach(s->productService.deleteProductImage(s.getImageFileName()));
 
-            productService.deleteProductImage(savedFileName);
+            }
+
             return "admin/productForm";
 
         } else {
-
+            product.setProductImages(savedImages);
             productRepository.save(product);
+            savedImages.stream()
+                    .forEach(s->productImageRepository.save(s));
             return "redirect:/admin/products";
         }
 
@@ -164,11 +222,8 @@ public class ProductController {
 
         if (productRepository.findById(productId).isPresent()) {
             Product product = productRepository.findById(productId).get();
+            logger.info("Images to delete: " + product.getProductImages().toString());
             productService.deleteProductImage(product.getImageFileName());
-//            File file = new File(rentierProperties.getUploadPathProductsForDelete() + product.getImageFileName());
-//            if (file.exists()) {
-//                file.delete();
-//            }
             productRepository.delete(product);
         }
